PHP 8.0 Windows support

+++ php/src/php/Operation.cpp
@@ -383,7 +383,7 @@ IcePHP::OperationI::getArgInfo(zend_internal_arg_info& arg, const ParamInfoPtr& 
     {
         zend_internal_arg_info ai[] =
         {
-            ZEND_ARG_ARRAY_INFO(pass_by_ref, 0, allow_null)
+            ZEND_ARG_ARRAY_INFO((uint32_t)pass_by_ref, (uint32_t)0, (uint32_t)allow_null)
         };
         arg = ai[0];
     }
@@ -391,7 +391,7 @@ IcePHP::OperationI::getArgInfo(zend_internal_arg_info& arg, const ParamInfoPtr& 
     {
         zend_internal_arg_info ai[] =
         {
-            ZEND_ARG_CALLABLE_INFO(pass_by_ref, 0, allow_null)
+            ZEND_ARG_CALLABLE_INFO((uint32_t)pass_by_ref, (uint32_t)0, allow_null)
         };
         arg = ai[0];
     }
